backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
PyMySQL==1.1.1
python-decouple==3.8
Pillow==10.1.0
django-filter==23.5
django-extensions==3.2.3
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
